# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BirthPath Support Inbox API",
    "version": "1.0.0",
    "description": "REST API for managing BirthPath support mailboxes, messages, drafts, folders, threads, search, and attachments.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "https://birthpath-support-inbox.god175.workers.dev"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/config": {
      "get": {
        "operationId": "getConfig",
        "summary": "Get configured domains and email addresses",
        "tags": [
          "Discovery"
        ],
        "responses": {
          "200": {
            "description": "Runtime configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Config"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/mailboxes": {
      "get": {
        "operationId": "listMailboxes",
        "summary": "List mailboxes",
        "tags": [
          "Mailboxes"
        ],
        "responses": {
          "200": {
            "description": "Mailboxes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Mailbox"
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "operationId": "createMailbox",
        "summary": "Create a mailbox",
        "tags": [
          "Mailboxes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMailboxRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mailbox created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mailbox"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/v1/mailboxes/{mailboxId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/MailboxId"
        }
      ],
      "get": {
        "operationId": "getMailbox",
        "summary": "Get a mailbox",
        "tags": [
          "Mailboxes"
        ],
        "responses": {
          "200": {
            "description": "Mailbox",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mailbox"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "operationId": "updateMailbox",
        "summary": "Replace mailbox settings",
        "tags": [
          "Mailboxes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMailboxRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mailbox updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mailbox"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "deleteMailbox",
        "summary": "Delete a mailbox",
        "tags": [
          "Mailboxes"
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/mailboxes/{mailboxId}/emails": {
      "parameters": [
        {
          "$ref": "#/components/parameters/MailboxId"
        }
      ],
      "get": {
        "operationId": "listEmails",
        "summary": "List mailbox emails",
        "tags": [
          "Emails"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/FolderQuery"
          },
          {
            "$ref": "#/components/parameters/ThreadIdQuery"
          },
          {
            "$ref": "#/components/parameters/ThreadedQuery"
          },
          {
            "$ref": "#/components/parameters/PageQuery"
          },
          {
            "$ref": "#/components/parameters/LimitQuery"
          },
          {
            "name": "sortColumn",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "date",
                "subject",
                "sender",
                "recipient",
                "read",
                "starred"
              ]
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Emails, or paginated emails when folder is supplied",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EmailMetadata"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/PaginatedEmails"
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "operationId": "sendEmail",
        "summary": "Send a new email",
        "tags": [
          "Emails"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendEmailRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Email accepted for delivery",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendEmailResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/v1/mailboxes/{mailboxId}/drafts": {
      "parameters": [
        {
          "$ref": "#/components/parameters/MailboxId"
        }
      ],
      "post": {
        "operationId": "createDraft",
        "summary": "Create or replace a draft",
        "tags": [
          "Emails"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/mailboxes/{mailboxId}/emails/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/MailboxId"
        },
        {
          "$ref": "#/components/parameters/EmailId"
        }
      ],
      "get": {
        "operationId": "getEmail",
        "summary": "Get an email",
        "tags": [
          "Emails"
        ],
        "responses": {
          "200": {
            "description": "Email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailFull"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "operationId": "updateEmail",
        "summary": "Update read and starred flags",
        "tags": [
          "Emails"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailFull"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "deleteEmail",
        "summary": "Delete an email",
        "tags": [
          "Emails"
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/mailboxes/{mailboxId}/emails/{id}/move": {
      "parameters": [
        {
          "$ref": "#/components/parameters/MailboxId"
        },
        {
          "$ref": "#/components/parameters/EmailId"
        }
      ],
      "post": {
        "operationId": "moveEmail",
        "summary": "Move an email to another folder",
        "tags": [
          "Emails"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveEmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email moved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/mailboxes/{mailboxId}/threads/{threadId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/MailboxId"
        },
        {
          "$ref": "#/components/parameters/ThreadId"
        }
      ],
      "get": {
        "operationId": "getThread",
        "summary": "Get all emails in a thread",
        "tags": [
          "Threads"
        ],
        "responses": {
          "200": {
            "description": "Thread emails",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmailFull"
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/mailboxes/{mailboxId}/threads/{threadId}/read": {
      "parameters": [
        {
          "$ref": "#/components/parameters/MailboxId"
        },
        {
          "$ref": "#/components/parameters/ThreadId"
        }
      ],
      "post": {
        "operationId": "markThreadRead",
        "summary": "Mark a thread as read",
        "tags": [
          "Threads"
        ],
        "responses": {
          "200": {
            "description": "Thread marked read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "Use the API key configured as the Worker's API_KEY secret."
      }
    },
...
    "responses": {
...
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NoContent": {
        "description": "Success with no response body"
      }
    },
    "schemas": {
...
      "Config": {
        "type": "object",
        "required": [
          "domains",
          "emailAddresses"
        ],
        "properties": {
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "emailAddresses": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "email"
            }
          }
        }
      },
...
      "Mailbox": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "settings"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "settings": {
            "$ref": "#/components/schemas/MailboxSettings"
          }
        }
      },
...
    }
  }
}